requirements.txt
accelerate
dotenv
einops
# for gemini api
google-genai
gradio
huggingface_hub
imageio
ipdb
json_repair
matplotlib
nltk
numpy
omegaconf
openai
opencv-python-headless
openpyxl
pandas
pillow
portalocker
protobuf
python-dotenv
qwen_vl_utils
requests
rich
sentencepiece
setuptools
sty
tabulate
tiktoken
timeout-decorator
timm
tqdm
typing_extensions
validators
xlsxwriter
